requirements.txt
fastapi
uvicorn
jinja2
pgmpy
python-telegram-bot
python-dotenv

